+++ spec/filters_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
-describe Weka::Clusterers do
+describe Weka::Filters do
   it_behaves_like 'class builder'
 
   %i[
-    Cobweb
-    Canopy
-    EM
-    FarthestFirst
-    FilteredClusterer
-    HierarchicalClusterer
-    MakeDensityBasedClusterer
-    SimpleKMeans
+    Filter
+    CheckSource
+    Filter
+    MultiFilter
+    RenameRelation
+    SimpleBatchFilter
+    SimpleFilter
+    SimpleStreamFilter
   ].each do |class_name|
     it "defines a class #{class_name}" do
       expect(described_class.const_defined?(class_name)).to be true